backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
netmiko
celery
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings
python-dotenv
nornir==3.4.1
nornir-napalm==0.4.0
nornir-netmiko==0.2.0
nornir-jinja2==0.2.0
jinja2==3.1.3
napalm==4.1.0
GitPython==3.1.41
